refactor: split Bedrock by model family and share provider request builders
